(improvement)(chat) Optimize the update parserInfo code and resolve compilation exceptions
