+++ chat/api/src/main/java/com/tencent/supersonic/chat/api/component/SemanticQuery.java
@@ -20,7 +20,5 @@ public interface SemanticQuery {
 
     SemanticParseInfo getParseInfo();
 
-    void updateParseInfo();
-
     void setParseInfo(SemanticParseInfo parseInfo);
 }
