+++ chat/core/src/main/java/com/tencent/supersonic/chat/service/ParseInfoService.java
@@ -1,0 +1,16 @@
+package com.tencent.supersonic.chat.service;
+
+import com.tencent.supersonic.chat.api.component.SemanticQuery;
+import com.tencent.supersonic.chat.api.pojo.SemanticParseInfo;
+import java.util.List;
+
+public interface ParseInfoService {
+
+    List<SemanticParseInfo> getTopCandidateParseInfo(List<SemanticParseInfo> selectedParses,
+            List<SemanticParseInfo> candidateParses);
+
+    List<SemanticParseInfo> sortParseInfo(List<SemanticQuery> semanticQueries);
+
+    void updateParseInfo(SemanticParseInfo parseInfo);
+
+}
