+++ chat/core/src/main/java/com/tencent/supersonic/chat/service/impl/ChatServiceImpl.java
@@ -223,10 +223,9 @@ public class ChatServiceImpl implements ChatService {
     }
 
     @Override
-    public List<ChatParseDO> batchAddParse(ChatContext chatCtx, QueryReq queryReq,
-            ParseResp parseResult,
-            List<SemanticParseInfo> candidateParses,
-            List<SemanticParseInfo> selectedParses) {
+    public List<ChatParseDO> batchAddParse(ChatContext chatCtx, QueryReq queryReq, ParseResp parseResult) {
+        List<SemanticParseInfo> candidateParses = parseResult.getCandidateParses();
+        List<SemanticParseInfo> selectedParses = parseResult.getSelectedParses();
         return chatQueryRepository.batchSaveParseInfo(chatCtx, queryReq, parseResult, candidateParses, selectedParses);
     }
